Tensorboard Learnings, Clean Up off Directory and Files
