+++ Deep Learning Basics with TensorFlow and Keras/TuT 4 Analyzing Models with Tensorboard.py
@@ -1,0 +1,55 @@
+import pickle
+import os
+import tensorflow as tf
+from keras.preprocessing.image import ImageDataGenerator
+from keras.datasets import cifar10
+from tensorflow.python.keras.models import Sequential
+from tensorflow.python.keras.layers import Dense, Dropout, Activation, Flatten
+from tensorflow.python.keras.layers import Conv2D, MaxPooling2D
+from tensorflow.python.keras.callbacks import TensorBoard
+import time
+import datetime
+
+log_dir = os.path.join(
+    "../logs",
+    "fit",
+    datetime.datetime.now().strftime("%Y%m%d-%H%M%S"),
+)
+
+NAME = "Cats-vs-dog-cnn-64x2-{}".format(int(time.time()))
+
+tensorboard = TensorBoard(log_dir="logs/{}".format(NAME))
+
+gpu_options = tf.compat.v1.GPUOptions(per_process_gpu_memory_fraction=0.333)
+sess = tf.compat.v1.Session(config=tf.compat.v1.ConfigProto(gpu_options=gpu_options))
+
+pickle_in = open("../X.pickle", "rb")
+X = pickle.load(pickle_in)
+
+pickle_in = open("../y.pickle", "rb")
+y = pickle.load(pickle_in)
+
+X = X/255.0
+
+model = Sequential()
+
+model.add(Conv2D(64, (3, 3), input_shape=X.shape[1:]))
+model.add(Activation('relu'))
+model.add(MaxPooling2D(pool_size=(2, 2)))
+
+model.add(Conv2D(64, (3, 3)))
+model.add(Activation('relu'))
+model.add(MaxPooling2D(pool_size=(2, 2)))
+
+model.add(Flatten())  # this converts our 3D features maps to 1D feature vectors
+
+model.add(Dense(64))
+model.add(Activation('relu'))
+
+model.add(Dense(1))
+model.add(Activation('sigmoid'))
+
+model.compile(loss='binary_crossentropy', optimizer='adam', metrics=['accuracy'])
+
+model.fit(X, y, batch_size=32, epochs=10, validation_split=0.3, callbacks=[tensorboard])
+
