--- Deep Learning TUT 3.py
@@ -1,39 +1,0 @@
-import pickle
-import numpy as np
-import tensorflow as tf
-from tensorflow.python.keras import layers
-from keras.datasets import cifar10
-from keras.preprocessing.image import ImageDataGenerator
-from tensorflow.python.keras import Sequential
-from tensorflow.python.keras.layers import Dense, Dropout, Activation, Flatten
-from tensorflow.python.keras.layers import Conv2D, MaxPooling2D
-
-pickle_in = open("X.pickle", "rb")
-X = pickle.load(pickle_in)
-
-pickle_in = open("y.pickle", "rb")
-y = pickle.load(pickle_in)
-
-X = X/255.0
-
-model = Sequential()
-
-model.add(Conv2D(64, (3, 3), input_shape=X.shape[1:]))
-model.add(Activation('relu'))
-model.add(MaxPooling2D(pool_size=(2, 2)))
-
-model.add(Conv2D(64, (3, 3)))
-model.add(Activation('relu'))
-model.add(MaxPooling2D(pool_size=(2, 2)))
-
-model.add(Flatten())  # this converts our 3D features maps to 1D feature vectors
-
-model.add(Dense(64))
-model.add(Activation('relu'))
-
-model.add(Dense(1))
-model.add(Activation('sigmoid'))
-
-model.compile(loss='binary_crossentropy', optimizer='adam', metrics=['accuracy'])
-
-model.fit(X, y, batch_size=32, epochs=3, validation_split=0.1)
