--- Deep Learning TuT 2.py
@@ -1,66 +1,0 @@
-import numpy as np
-import matplotlib.pyplot as plt
-import os
-import cv2
-from tqdm import tqdm
-import random
-import pickle
-
-"""Creating the Dataset"""
-
-DATADIR = "D:/Code/PetImages"
-
-CATEGORIES = ["Dog", "Cat"]
-IMG_SIZE = 50
-training_data = []
-
-def create_training_data():
-    for category in CATEGORIES:  # do dogs and cats
-
-        path = os.path.join(DATADIR, category)  # create path to dogs and cats
-        class_num = CATEGORIES.index(category)  # get the classification (0 or 1). 0=dog 1=cat
-
-        for img in tqdm(os.listdir(path)):  # iterate over each image per dogs and cats
-            try:
-                img_array = cv2.imread(os.path.join(path, img), cv2.IMREAD_GRAYSCALE)  # convert to array
-                new_array = cv2.resize(img_array, (IMG_SIZE, IMG_SIZE))  # resize to normalize data size
-                training_data.append([new_array, class_num])  # add this to our training data
-            except Exception as e:  # in the interest in keeping the output clean...
-                pass
-            except OSError as e:
-               print("OSError Bad img most likely", e, os.path.join(path, img))
-            except Exception as e:
-               print("general exception", e, os.path.join(path, img))
-
-create_training_data()
-random.shuffle(training_data)  # so no 12k cats into 12k dogs straight
-
-"""Model"""
-
-X = []
-y = []
-
-for features,label in training_data:
-    X.append(features)
-    y.append(label)
-
-print(X[0].reshape(-1, IMG_SIZE, IMG_SIZE, 1))
-
-X = np.array(X).reshape(-1, IMG_SIZE, IMG_SIZE, 1)
-y = np.array(y)
-"""Saving the model for later"""
-
-pickle_out = open("X.pickle", "wb")
-pickle.dump(X, pickle_out)
-pickle_out.close()
-
-pickle_out = open("y.pickle", "wb")
-pickle.dump(y, pickle_out)
-pickle_out.close()
-
-"""Loading the model """
-pickle_in = open("X.pickle", "rb")
-X = pickle.load(pickle_in)
-
-pickle_in = open("y.pickle", "rb")
-y = pickle.load(pickle_in)
